Victory Lap
My main change in this push is the addition of animations at the end of
each level, as opposed to the instant freezing of the screen. Now you
have 3 seconds to admire your victory as you smoothly pull in to dock
with the escape pod, or agonize over your untimely demise as the massive
explosion that was once your ship fades away.
I also improved hitboxes, so those are more accurate now.