Correct Physics
Okay, I was kind of afraid this would happen; I'm not actually sure how
to get back to Master, so there is an off-chance I am just forever stuck
in Alternate-Dimension now. Oh, well.
In other news, I finally corrected my physics! It may not look like
much, but there were a lot of calculus and algebreaic calculations on my
side of this commit. You may now rely on the finality and correctness of
my physics.