Slow Down
I implemented a secondary means of death accomplished when one attempts
to fly at the speed of light, which triggers a new defeat sound effect.
I might still work on the animation for that, but it works for now. I
also made collisions slow you down, making it more difficult to power
through clumps of asteroids.